Allow for returning only figure by default, and add curves/clusters dictionary.
By default, the user will no longer need to unpack more than one value (as was the case before, when the computed dendrogram traces were returned along with the figure). Additionally, if the user wishes to make use of hover/click data, a dictionary can optionally be returned that translates 'curve number' in the figure object to a cluster (e.g., row cluster 1, col cluster 2).