Allow for optional trace computation; return values instead of modifying them.
In the event that the entire dendrogram and heatmap do not need to be fully recalculated (e.g., if one is showing/hiding row/column labels or adding row/column group markers), use the heatmap, row labels, and column labels from the existing clustergram.
Rename the function _dendrogram_traces() to _compute_clustered_data, to highlight the fact that it additionally reorders the data and row/column labels to fit the dendrogram leaves after clustering.
Instead of directly modifying properties of the clustergram object in _compute_clustered_data(), return values from _compute_clustered_data and directly modify the properties in the figure() function to avoid hidden side-effects.