refactor: deploy monorepo correctly to playground
This change refactors the demo python files into actual standalone applications. While this does increase the overhead around copy-pasting files and keeping things in sync, it decreases the amount of work necessary to reason about what files go where.

As a side-effect, it also allows us to switch to a much faster deploy process because we don't need to clone the entire repository.

Finally, the GA tag is updated to the latest version.