Fix error in Manhattan plot for single chromosome
When attempting to use ManhattanPlot on data containing a single
chromosome, a runtime error is encountered. Seems it was previously
reported in #486 but the cause was not determined. From looking at the
code, it seems like the `name` parameter was incorrectly included inside
the `marker` dict, when it should have been a direct kwarg.